trying to get port to work

+++ config/Database.php
@@ -15,7 +15,7 @@
             $this->password = getenv('PASSWORD');
             $this->db_name = getenv('DBNAME');
             $this->host = getenv('HOST');
-            $this->port = getenv('PORT');
+            $this->port = (int) getenv('PORT');
         }
 
         // DB connectg
